+++ lua/ecovim/config/plugins.lua
@@ -84,6 +84,7 @@ if settings.enable_terminal then
       "IPython",
       "Lazyman",
       "Lazyconf",
+      "Lazyplug",
       "Htop",
     },
     event = "VimEnter",