+++ lua/onno/plugins/ui.lua
@@ -319,6 +319,7 @@ if settings.enable_terminal then
       "IPython",
       "Lazyman",
       "Lazyconf",
+      "Lazyplug",
       "Htop",
     },
     event = "VimEnter",